@@ -11,28 +11,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.0.0",
-    "@angular/common": "^14.0.0",
-    "@angular/compiler": "^14.0.0",
-    "@angular/core": "^14.0.0",
-    "@angular/forms": "^14.0.0",
-    "@angular/platform-browser": "^14.0.0",
-    "@angular/platform-browser-dynamic": "^14.0.0",
-    "@angular/router": "^14.0.0",
-    "bootstrap": "^5.1.3",
-    "jquery": "^3.6.0",
-    "keycloak-js": "^18.0.0",
-    "ngx-spinner": "^13.1.1",
+    "@angular/animations": "^15.1.3",
+    "@angular/common": "^15.1.3",
+    "@angular/compiler": "^15.1.3",
+    "@angular/core": "^15.1.3",
+    "@angular/forms": "^15.1.3",
+    "@angular/platform-browser": "^15.1.3",
+    "@angular/platform-browser-dynamic": "^15.1.3",
+    "@angular/router": "^15.1.3",
+    "bootstrap": "^5.2.3",
+    "jquery": "^3.6.3",
+    "keycloak-js": "^20.0.3",
+    "ngx-spinner": "^15.0.1",
     "popper.js": "^1.16.1",
-    "rxjs": "~7.5.5",
-    "tslib": "^2.3.1",
-    "zone.js": "~0.11.5"
+    "prettier": "^2.8.3",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.5.0",
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "14.0.0",
-    "@angular/cli": "^14.0.0",
-    "@angular/compiler-cli": "^14.0.0",
-    "@angular/language-service": "^14.0.0",
+    "@angular-devkit/build-angular": "15.1.4",
+    "@angular/cli": "^15.1.4",
+    "@angular/compiler-cli": "^15.1.3",
+    "@angular/language-service": "^15.1.3",
     "@types/jasmine": "~3.5.14",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "^17.0.39",
@@ -47,6 +48,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~9.0.0",
     "tslint": "~6.1.0",
-    "typescript": "^4.7.3"
+    "typescript": "^4.9.5"
   }
 }
