@@ -1,12 +1,10 @@
-import {Component} from "@angular/core";
+import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+	selector: 'app-root',
+	templateUrl: './app.component.html',
+	styleUrls: ['./app.component.css'],
 })
-export class AppComponent
-{
-  title = 'KeyCloak Demo';
-
+export class AppComponent {
+	title = 'KeyCloak Demo';
 }
