@@ -1,12 +1,12 @@
-import {TestBed} from "@angular/core/testing";
+import { TestBed } from '@angular/core/testing';
 
-import {EmployeeService} from "./employee.service";
+import { EmployeeService } from './employee.service';
 
 describe('EmployeeService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+	beforeEach(() => TestBed.configureTestingModule({}));
 
-  it('should be created', () => {
-    const service: EmployeeService = TestBed.get(EmployeeService);
-    expect(service).toBeTruthy();
-  });
+	it('should be created', () => {
+		const service: EmployeeService = TestBed.get(EmployeeService);
+		expect(service).toBeTruthy();
+	});
 });
