@@ -1,8 +1,7 @@
-import {UserProfile} from "./user-profile";
+import { UserProfile } from './user-profile';
 
-export class Employee extends UserProfile
-{
-  employeeId: number;
-  location: string;
-  salary: number;
+export class Employee extends UserProfile {
+	employeeId: number;
+	location: string;
+	salary: number;
 }
