@@ -1,8 +1,7 @@
-export class UserProfile
-{
-  id: number;
-  firstName: string;
-  lastName: string;
-  email: string;
-  phone: string;
+export class UserProfile {
+	id: number;
+	firstName: string;
+	lastName: string;
+	email: string;
+	phone: string;
 }
