@@ -1,25 +1,24 @@
-import {async, ComponentFixture, TestBed} from "@angular/core/testing";
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import {EmployeeProjectComponent} from "./employee-project.component";
+import { EmployeeProjectComponent } from './employee-project.component';
 
 describe('EmployeeProjectComponent', () => {
-  let component: EmployeeProjectComponent;
-  let fixture: ComponentFixture<EmployeeProjectComponent>;
+	let component: EmployeeProjectComponent;
+	let fixture: ComponentFixture<EmployeeProjectComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ EmployeeProjectComponent ]
-    })
-    .compileComponents();
-  }));
+	beforeEach(async(() => {
+		TestBed.configureTestingModule({
+			declarations: [EmployeeProjectComponent],
+		}).compileComponents();
+	}));
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(EmployeeProjectComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+	beforeEach(() => {
+		fixture = TestBed.createComponent(EmployeeProjectComponent);
+		component = fixture.componentInstance;
+		fixture.detectChanges();
+	});
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
+	it('should create', () => {
+		expect(component).toBeTruthy();
+	});
 });
