@@ -1,12 +1,12 @@
-import {TestBed} from "@angular/core/testing";
+import { TestBed } from '@angular/core/testing';
 
-import {EmployeeProjectService} from "./employee-project.service";
+import { EmployeeProjectService } from './employee-project.service';
 
 describe('EmployeeProjectService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+	beforeEach(() => TestBed.configureTestingModule({}));
 
-  it('should be created', () => {
-    const service: EmployeeProjectService = TestBed.get(EmployeeProjectService);
-    expect(service).toBeTruthy();
-  });
+	it('should be created', () => {
+		const service: EmployeeProjectService = TestBed.get(EmployeeProjectService);
+		expect(service).toBeTruthy();
+	});
 });
