@@ -1,12 +1,12 @@
-import {TestBed} from "@angular/core/testing";
+import { TestBed } from '@angular/core/testing';
 
-import {KeycloakService} from "./keycloak.service";
+import { KeycloakService } from './keycloak.service';
 
 describe('KeycloakService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+	beforeEach(() => TestBed.configureTestingModule({}));
 
-  it('should be created', () => {
-    const service: KeycloakService = TestBed.get(KeycloakService);
-    expect(service).toBeTruthy();
-  });
+	it('should be created', () => {
+		const service: KeycloakService = TestBed.get(KeycloakService);
+		expect(service).toBeTruthy();
+	});
 });
